Reorganize api into files
